Meta engine API support.
