+++ composer.json
@@ -15,7 +15,7 @@
   ],
   "require" : {
     "php" : "^5.6|^7.0",
-    "elastic/openapi-codegen" : "^1.0.3",
+    "elastic/openapi-codegen" : "^1.0.4",
     "psr/log" : "^1.0."
   },
   "require-dev" : {
