+++ resources/api/templates/README.mustache
@@ -1,0 +1,16 @@
+{{>readme_title}}
+{{>readme_content}}
+{{>readme_install}}
+{{>readme_usage}}
+### Clients methods
+
+Method      | Description | Documentation
+------------|-------------|--------------
+**`createEngine`**| Creates a new engine.<br/> <br/> **Parameters :** <br />  - `$name` (required) <br />   - `$language`<br/>|[Endpoint Documentation](https://swiftype.com/documentation/app-search/api/engines#create)
+**`createMetaEngine`**| Creates a new meta engine.<br/> <br/> **Parameters :** <br />  - `$name` (required) <br />   - `$sourceEngines` (required)<br/>|[Endpoint Documentation](https://swiftype.com/documentation/app-search/api/engines#create)
+{{#apiInfo}}{{#apis}}{{#operations}}{{#operation}}{{#vendorExtensions.x-add-to-documentation}}**`{{operationId}}`**| {{#summary}}{{.}}{{/summary}}{{#allParams.0}}<br/> <br/> **Parameters :** <br />{{#allParams}}  - `${{paramName}}`{{#required}} (required) {{/required}}{{#hasMore}}<br /> {{/hasMore}}{{/allParams}}{{/allParams.0}}<br/>{{#externalDocs}}|[Endpoint Documentation]({{url}}){{/externalDocs}}
+{{/vendorExtensions.x-add-to-documentation}}{{/operation}}{{/operations}}{{/apis}}{{/apiInfo}}
+{{>readme_dev}}
+{{>readme_faq}}
+{{>readme_contrib}}
+{{>readme_licence}}
